Move npc/ipset package to net/ipset
It's going to be used by net/bridge.go.
